# Conflicts:
#	src/main/front_end/package-lock.json
#	src/main/front_end/src/actions/AuthActions.js
#	src/main/front_end/src/actions/ChooseStudentRoomActions.js
#	src/main/front_end/src/actions/OpCodeTypes.js
#	src/main/front_end/src/actions/StudentActions.js
#	src/main/front_end/src/actions/TeacherActions.js
#	src/main/front_end/src/actions/WatchMessagesActions.js
#	src/main/front_end/src/actions/WatchProfileActions.js
#	src/main/front_end/src/actions/types.js
#	src/main/front_end/src/api/API.js
#	src/main/front_end/src/api/APIPaths.js
#	src/main/front_end/src/components/screens/ITScreen.js
#	src/main/front_end/src/components/screens/StudentScreen.js
#	src/main/front_end/src/components/screens/TeacherScreen.js
#	src/main/front_end/src/components/screens/index.js
#	src/main/front_end/src/index.js
#	src/main/front_end/src/locale/locale_heb.js
#	src/main/front_end/src/navigation/NavPaths.js
#	src/main/front_end/src/reducers/SolveStoryReducer.js
#	src/main/front_end/src/reducers/WatchSuggestionReducer.js
#	src/main/front_end/src/reducers/index.js
#	src/main/front_end/yarn.lock
#	src/main/java/DataAPI/MissionData.java
#	src/main/java/DataAPI/OpCode.java
#	src/main/java/DataAPI/RoomDetailsData.java
#	src/main/java/Utils/Utils.java
#	src/main/java/missions/room/Communications/Controllers/ITController.java
#	src/main/java/missions/room/Communications/Publisher/SinglePublisher.java
#	src/main/java/missions/room/Communications/RealTime/WebSocketConfig.java
#	src/main/java/missions/room/Domain/Classroom.java
#	src/main/java/missions/room/Domain/Notifications/NonPersistenceNotification.java
#	src/main/java/missions/room/Domain/Ram.java
#	src/main/java/missions/room/Domain/Rooms/ClassroomRoom.java
#	src/main/java/missions/room/Domain/Rooms/Room.java
#	src/main/java/missions/room/Domain/Rooms/StudentRoom.java
#	src/main/java/missions/room/Domain/Users/SchoolUser.java
#	src/main/java/missions/room/Domain/Users/Student.java
#	src/main/java/missions/room/Domain/Users/Teacher.java
#	src/main/java/missions/room/Domain/Users/User.java
#	src/main/java/missions/room/Domain/missions/KnownAnswerMission.java
#	src/main/java/missions/room/Domain/missions/Mission.java
#	src/main/java/missions/room/Managers/ManagerRoomStudent.java
#	src/main/java/missions/room/Managers/ProfileMessagesManager.java
#	src/main/java/missions/room/Managers/RoomManager.java
#	src/main/java/missions/room/Managers/SuggestionManager.java
#	src/main/java/missions/room/Repo/OpenAnswerRepo.java
#	src/main/java/missions/room/Repo/SuggestionRepo.java
#	src/main/java/missions/room/Repo/UserRepo.java
#	src/main/java/missions/room/RoomApplication.java
#	src/main/java/missions/room/Service/RoomService.java
#	src/main/java/missions/room/Service/ServiceAPI.java
#	src/main/java/missions/room/Service/StudentRoomService.java
#	src/main/resources/application.properties
#	src/test/java/Data/Data.java
#	src/test/java/Data/DataConstants.java
#	src/test/java/Data/DataGenerator.java
#	src/test/java/DomainMocks/PublisherMock.java
#	src/test/java/missions/room/ManagerRoomStudentTests/ManagerRoomStudentAllStubs.java
#	src/test/java/missions/room/ManagerRoomStudentTests/ManagerRoomStudentRealRam.java
#	src/test/java/missions/room/ManagerRoomStudentTests/ManagerRoomStudentRealRamRealClassRoomRealRamStudentClassroomClassGroupRepo.java
#	src/test/java/missions/room/ManagerRoomStudentTests/ManagerRoomStudentRealRoomRamStudentClassroomClassGroupRepo.java
#	src/test/java/missions/room/SuggestionManagerTests/SuggestionManagerTestsAllStubs.java
#	src/test/resources/init_users_db.sql

+++ src/main/java/missions/room/Managers/RoomManager.java
@@ -249,7 +249,11 @@ public class RoomManager extends TeacherManager {
         if(room.getConnectedUsersAliases().size()>0){
             return new Response<>(false, OpCode.CONNECTED_STUDENTS);
         }
-        //TODO don't close room if someone is connected to it
+        //TODO sync!!! on the room
+        OpCode response=ram.connectToRoom(roomId,ram.getAlias(apiKey));
+        if(response!=OpCode.Teacher){
+            return new Response<>(false,response);
+        }
         return roomRepo.deleteRoom(room);
     }
